+++ constants/index.js
@@ -1,0 +1,6 @@
+const URLS = {
+  certificate:
+    "https://firebasestorage.googleapis.com/v0/b/servicedogschool-e0fc7.appspot.com/o/SA-2809131-C.pdf?alt=media&token=a314d2a9-a263-45c4-9b8a-9f5f933cdae7",
+
+};
+module.exports = { URLS };
